The linear analysis is functionally complete. Have to integrate into our running framework, and pray that there aren't too many bugs when it runs on the ubuntu repos.
What this does:
If the dynsyms in the elf lists functions, it will attempt to break the
text section into functions and anlyze them as functions. This is
important to ensure that we are able to count instructuions in functions in
libraries independently from each other (otherwise we end up over
counting instructions.

However, in the event there is no function data (or no size available),
we treat the whole .text section as a single large binary chunk and
disassemble it from start to end.

Hopefully, we will still be able to mangle calls, otherwise we won't be
able to do the linking analysis. Let's see.

The linear method may include junk instructions (data embedded in the
text section, etc) but this is favorable to  the poor coverage we get
when we attempt to do control flow analysis. Previous attempts at
building a control flow engine have yielded results as bad as ~2% coverage.