Tweaked how we store register values, and introduced the notion of concreteness.
If a value is 'mov'-ed into a register, that value is concrete, until
we hit a control flow boundary. After that some registers will remain
concrete (because they are callee-saved), however, others will need to
be invalidated, and only become conrete again if they are repopulated
using a pop or a mov.

Partially incomplete, as pop and push aren't being specially tracked
yet.